editor table in preface: https://github.com/metanorma/metanorma-itu/issues/336
