+++ lib/isodoc/itu/i18n-es.yaml
@@ -60,6 +60,8 @@ br_full: Oficina de Radiocomunicaciones<br/>de la UIT
 bdt_full: <br/>Oficina de Desarrollo<br/>de las Telecomunicaciones de la UIT
 prepub: Versión prepublicada
 draft_new: Borrador de nueva %
+email: E-mail
+editor_full: editor
 doctype_dict:
   resolution: Resolución
   recommendation: Recomendación
@@ -87,3 +89,14 @@ doctype_abbrev:
   technical-report: TR
   joint-itu-iso-iec: UIT/ISO/IEC
   service-publication: 
+inflection:
+  editor:
+    masc:
+      sg: editor
+      pl: editores
+    fem:
+      sg: editora
+      pl: editoras
+    nb:
+      sg: editor@
+      pl: editor@s
