+++ lib/isodoc/itu/i18n-zh-Hans.yaml
@@ -53,6 +53,8 @@ br_full: 国际电信联盟<br/>电信发展部门
 bdt_full: 国际电信联盟<br/>电信发展局
 prepub: 预发布版本
 draft_new: 新%草案
+email: 电子邮件
+editor_full: 编辑
 doctype_dict:
   resolution: 决议
   recommendation: 建议书
